--- a/pkg/store/tsdb.go
+++ b/pkg/store/tsdb.go
@@ -5,6 +5,7 @@ package store
 
 import (
 	"context"
+	"io"
 	"math"
 	"sort"
 
@@ -13,6 +14,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
+	storetestutil "github.com/thanos-io/thanos/pkg/store/storepb/testutil"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -54,7 +56,7 @@ func NewTSDBStore(logger log.Logger, _ prometheus.Registerer, db TSDBReader, com
 		db:               db,
 		component:        component,
 		externalLabels:   externalLabels,
-		maxBytesPerFrame: 1024 * 1024, // 1MB as recommended by gRPC.
+		maxBytesPerFrame: storetestutil.RemoteReadFrameLimit,
 	}
 }
 
@@ -89,6 +91,12 @@ func (s *TSDBStore) Info(_ context.Context, _ *storepb.InfoRequest) (*storepb.In
 	return res, nil
 }
 
+// CloseDelegator allows to delegate close (releasing resources used by request to the server).
+// This is useful when we invoke StoreAPI within another StoreAPI and results are ephemeral until copied.
+type CloseDelegator interface {
+	Delegate(io.Closer)
+}
+
 // Series returns all series for a requested time range and label matcher. The returned data may
 // exceed the requested time bounds.
 func (s *TSDBStore) Series(r *storepb.SeriesRequest, srv storepb.Store_SeriesServer) error {
@@ -114,30 +122,33 @@ func (s *TSDBStore) Series(r *storepb.SeriesRequest, srv storepb.Store_SeriesSer
 	if err != nil {
 		return status.Error(codes.Internal, err.Error())
 	}
-	defer runutil.CloseWithLogOnErr(s.logger, q, "close tsdb querier series")
 
-	var (
-		set        = q.Select(false, nil, matchers...)
-		respSeries storepb.Series
-	)
+	if cd, ok := srv.(CloseDelegator); ok {
+		cd.Delegate(q)
+	} else {
+		defer runutil.CloseWithLogOnErr(s.logger, q, "close tsdb chunk querier series")
+	}
+
+	set := q.Select(false, nil, matchers...)
 
 	// Stream at most one series per frame; series may be split over multiple frames according to maxBytesInFrame.
 	for set.Next() {
 		series := set.At()
-		respSeries.Labels = s.translateAndExtendLabels(series.Labels(), s.externalLabels)
-		respSeries.Chunks = respSeries.Chunks[:0]
+		seriesLabels := storepb.Series{Labels: s.translateAndExtendLabels(series.Labels(), s.externalLabels)}
 		if r.SkipChunks {
-			if err := srv.Send(storepb.NewSeriesResponse(&respSeries)); err != nil {
+			if err := srv.Send(storepb.NewSeriesResponse(&seriesLabels)); err != nil {
 				return status.Error(codes.Aborted, err.Error())
 			}
 			continue
 		}
 
-		frameBytesLeft := s.maxBytesPerFrame
-		for _, lbl := range respSeries.Labels {
-			frameBytesLeft -= lbl.Size()
+		bytesLeftForChunks := s.maxBytesPerFrame
+		for _, lbl := range seriesLabels.Labels {
+			bytesLeftForChunks -= lbl.Size()
 		}
+		frameBytesLeft := bytesLeftForChunks
 
+		seriesChunks := []storepb.AggrChunk{}
 		chIter := series.Iterator()
 		isNext := chIter.Next()
 		for isNext {
@@ -146,15 +157,16 @@ func (s *TSDBStore) Series(r *storepb.SeriesRequest, srv storepb.Store_SeriesSer
 				return status.Errorf(codes.Internal, "TSDBStore: found not populated chunk returned by SeriesSet at ref: %v", chk.Ref)
 			}
 
-			respSeries.Chunks = append(respSeries.Chunks, storepb.AggrChunk{
+			c := storepb.AggrChunk{
 				MinTime: chk.MinTime,
 				MaxTime: chk.MaxTime,
 				Raw: &storepb.Chunk{
 					Type: storepb.Chunk_Encoding(chk.Chunk.Encoding() - 1), // Proto chunk encoding is one off to TSDB one.
 					Data: chk.Chunk.Bytes(),
 				},
-			})
-			frameBytesLeft -= respSeries.Chunks[len(respSeries.Chunks)-1].Size()
+			}
+			frameBytesLeft -= c.Size()
+			seriesChunks = append(seriesChunks, c)
 
 			// We are fine with minor inaccuracy of max bytes per frame. The inaccuracy will be max of full chunk size.
 			isNext = chIter.Next()
@@ -162,10 +174,11 @@ func (s *TSDBStore) Series(r *storepb.SeriesRequest, srv storepb.Store_SeriesSer
 				continue
 			}
 
-			if err := srv.Send(storepb.NewSeriesResponse(&respSeries)); err != nil {
+			if err := srv.Send(storepb.NewSeriesResponse(&storepb.Series{Labels: seriesLabels.Labels, Chunks: seriesChunks})); err != nil {
 				return status.Error(codes.Aborted, err.Error())
 			}
-			respSeries.Chunks = respSeries.Chunks[:0]
+			frameBytesLeft = bytesLeftForChunks
+			seriesChunks = make([]storepb.AggrChunk, 0, len(seriesChunks))
 		}
 		if err := chIter.Err(); err != nil {
 			return status.Error(codes.Internal, errors.Wrap(err, "chunk iter").Error())