@bwplotka 2020-08-24
Looking on different frame limits it's really close... However actually now it shows 1MB being bit better one...

```
bash ../thanos/_dev/benchmarks/cmp.sh noframelimit 1mblimit
Comparing '/home/bwplotka/Repos/thanos/_dev/benchmarks/results/BenchmarkTSDBStoreSeries/noframelimit/2020-08-24-T17-39-13.bench.out' with (new) '/home/bwplotka/Repos/thanos/_dev/benchmarks/results/BenchmarkTSDBStoreSeries/1mblimit/2020-08-24-T17-31-18.bench.out'
name                                                                                                    old time/op    new time/op    delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       6.24s ± 0%     6.73s ± 0%   +7.79%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      804ms ± 0%     684ms ± 0%  -15.02%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     426ms ± 0%     390ms ± 0%   -8.50%

name                                                                                                    old alloc/op   new alloc/op   delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12      2.16GB ± 0%    2.16GB ± 0%   -0.01%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      363MB ± 0%     362MB ± 0%   -0.25%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     601MB ± 0%     591MB ± 0%   -1.62%

name                                                                                                    old allocs/op  new allocs/op  delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       43.0M ± 0%     43.0M ± 0%   -0.00%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      4.43M ± 0%     4.43M ± 0%   -0.01%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12      518k ± 0%      519k ± 0%   +0.20%
```