@bwplotka 2020-08-24
Another one: 

```
name                                                                                                    old time/op    new time/op    delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       6.80s ± 0%     6.73s ± 0%   -1.08%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      970ms ± 0%     684ms ± 0%  -29.54%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     461ms ± 0%     390ms ± 0%  -15.55%

name                                                                                                    old alloc/op   new alloc/op   delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12      2.16GB ± 0%    2.16GB ± 0%   -0.00%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      365MB ± 0%     362MB ± 0%   -0.62%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     603MB ± 0%     591MB ± 0%   -1.93%

name                                                                                                    old allocs/op  new allocs/op  delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       43.0M ± 0%     43.0M ± 0%   -0.00%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      4.43M ± 0%     4.43M ± 0%   -0.02%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12      518k ± 0%      519k ± 0%   +0.19%
```