@bwplotka 2020-08-25
OK, extensive tests shows really not much difference, so sticking to 1MB limit as it was previously (: 

```
Comparing '/home/bwplotka/Repos/thanos/_dev/benchmarks/results/BenchmarkTSDBStoreSeries/noframelimit/2020-08-25-T12-07-03.bench.out' with (new) '/home/bwplotka/Repos/thanos/_dev/benchmarks/results/BenchmarkTSDBStoreSeries/1mblimit/2020-08-25-T11-45-07.bench.out'
name                                                                                                    old time/op    new time/op    delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       6.28s ± 0%     6.04s ± 0%  -3.96%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      814ms ± 0%     790ms ± 0%  -2.97%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     425ms ± 0%     427ms ± 0%  +0.48%

name                                                                                                    old alloc/op   new alloc/op   delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12      2.15GB ± 0%    2.15GB ± 0%  -0.00%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      365MB ± 0%     365MB ± 0%  -0.20%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12     612MB ± 0%     599MB ± 0%  -2.01%

name                                                                                                    old allocs/op  new allocs/op  delta
TSDBStoreSeries/1000000SeriesWith1Samples/3_blocks_and_one_WAL_with_1_samples,_250000_series_each-12       43.0M ± 0%     43.0M ± 0%  +0.00%
TSDBStoreSeries/100000SeriesWith100Samples/3_blocks_and_one_WAL_with_25_samples,_25000_series_each-12      4.43M ± 0%     4.43M ± 0%  -0.01%
TSDBStoreSeries/1SeriesWith10000000Samples/3_blocks_and_one_WAL_with_2500000_samples,_1_series_each-12      519k ± 0%      520k ± 0%  +0.18%
```