@bwplotka 2020-08-25
This is without marshalling - will do marshalling later on